requirements-core.txt
# Core dependencies for modular video generation system
google-generativeai>=0.3.0    # Gemini API for image generation and vision
pillow>=10.0.0                # Image processing
numpy>=1.24.0                 # Array operations
python-dotenv>=1.0.0          # .env file support for API keys
pyyaml>=6.0.0                 # YAML support for configuration

# Optional but recommended
scikit-learn>=1.3.0           # K-means clustering for color extraction
scipy>=1.11.0                 # Advanced image processing

# Development dependencies (optional)
pytest>=7.4.0                 # Testing framework
black>=23.0.0                 # Code formatting
flake8>=6.0.0                 # Linting
mypy>=1.0.0                   # Type checking
